Light和Transform的基础

+++ GamApp/GameScene/TestScene.cpp
@@ -55,31 +55,31 @@ void TestScene::OnBeginFrame()
 	D3DXVECTOR3 move(0, 0, 0);
 	if (KEYDOWN('W'))
 	{
-		move = D3DXVECTOR3(-1, 0, 0) * CameraParam::speed*dTime;
+		move = D3DXVECTOR3(-1, 0, 0) * CameraParam::speed*(float)dTime;
 	}
 
 	if (KEYDOWN('S'))
 	{
-		move = D3DXVECTOR3(1, 0, 0) * CameraParam::speed*dTime;
+		move = D3DXVECTOR3(1, 0, 0) * CameraParam::speed*(float)dTime;
 	}
 
 	if (KEYDOWN('A'))
 	{
-		move = D3DXVECTOR3(0, 0, -1) * CameraParam::speed*dTime;
+		move = D3DXVECTOR3(0, 0, -1) * CameraParam::speed*(float)dTime;
 	}
 
 	if (KEYDOWN('D'))
 	{
-		move = D3DXVECTOR3(0, 0, 1) * CameraParam::speed*dTime;
+		move = D3DXVECTOR3(0, 0, 1) * CameraParam::speed*(float)dTime;
 	}
 	if (KEYDOWN('Q'))
 	{
-		move = D3DXVECTOR3(0, 1, 0) * CameraParam::speed*dTime;
+		move = D3DXVECTOR3(0, 1, 0) * CameraParam::speed*(float)dTime;
 	}
 
 	if (KEYDOWN('E'))
 	{
-		move = D3DXVECTOR3(0, -1, 0) * CameraParam::speed*dTime;
+		move = D3DXVECTOR3(0, -1, 0) * CameraParam::speed*(float)dTime;
 	}
 	D3DXMatrixTranslation(&moveMat, move.x, move.y, move.z);
 	cW = cW * moveMat;
